pyproject.toml
[tool.poetry]
name = "hrbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
telebot = "^0.0.5"
python-dotenv = "^1.0.1"
langchain = "^0.3.11"
openai = "^1.57.2"
colorlog = "^6.9.0"
langchain-community = "^0.3.11"
langchain-openai = "^0.2.12"
pypdf2 = "^3.0.1"
retry = "^0.9.2"
supabase = "^2.13.0"
asyncio = "^3.4.3"
langgraph = "^0.2.73"
psycopg = {extras = ["binary", "pool"], version = "^3.2.4"}
langchain-postgres = "^0.0.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
